Check if test infrastructure is present and refers to the current package

Check if test infrastructure is present and refers to the current package

Usage

check_test_infra(
  path_infra = fs::path_wd("tests", "tinytest.R"),
  signal = c("error", "warning", "message", "quiet")
)

Arguments

path_infra

character string with a path to the file which determines the test infrastructure, see Details.

signal

character string indicating the type of signal to be used: "error" to throw an error, "warning" to issue a warning, "message" to show a message, or "quiet" to be quiet.

Value

A character vector containing a character string with the text of the signal, which is "" if no signal is emitted.

Details

The default path_infra points to file tinytest.R in directory tests of the current working directory, which belongs to the test infrastructure of tinytest. To use the test infrastructure of testthat, use path_infra = fs::path_wd("tests", "testthat.R").

The signal indicated by argument signal is emitted:

if the package indicated in path_infra is not among the dependencies

if the file that determines the used testing infrastructure does not exist

if the file that determines the used testing infrastructure does not refer to the package in the current working directory