Get test files

Get test files, emitting a signal if test files do not have corresponding function files or vice versa.

Usage

get_test_files(
  testdir,
  pattern = "^test_|^test-",
  ignore_case = TRUE,
  signal = c("error", "warning", "message", "quiet")
)

Arguments

testdir

character string containing a valid path to the directory containing tests.

pattern

character string containing a regular expression used to select test files.

ignore_case

TRUE or FALSE: ignore case when matching filenames to pattern?

signal

character string indicating the type of signal to be used: "error" to throw an error, "warning" to issue a warning, "message" to show a message, or "quiet" to be quiet.

Value

A character vector containing the found test files, which is character(0) if no test files are found, with attribute "info" containing a character string with the text of the signal, which is "" if no signal is emitted.

Details

The signal indicated by argument signal is emitted if directory testdir:

does not exist

contains files that are ignored because they do not start with pattern pattern, are not R files, or are template files created by tinytest

contain test files without corresponding function files in folder R

does not contain any test files

A character string with the text of this signal is also present as attribute "info" of the returned value.

See also